+++ ChitChatClient/Windows/ChatWindow.xaml.cs
@@ -1,10 +1,12 @@
 ﻿using ChitChatClient.Models;
+using ChitChatClient.Services;
 using ChitChatClient.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Timers;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
@@ -13,6 +15,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using System.Windows.Threading;
 
 namespace ChitChatClient.Windows
 {
@@ -22,26 +25,41 @@ namespace ChitChatClient.Windows
 	public partial class ChatWindow : Window
 	{
 		public ChatInterface chatInterface = new();
+
+		private static DispatcherTimer? userRefreshTimer;
+
+		private readonly ServerService serverService = ServerService.Instance;
+
 		public ChatWindow()
 		{
 			InitializeComponent();
 			DataContext = chatInterface;
+
+			// Initializing user refresh timer
+			userRefreshTimer = new DispatcherTimer();
+
+			userRefreshTimer.Tick += new EventHandler(RefreshServerUsersTimed);
+			userRefreshTimer.Interval = new TimeSpan(0, 0, 5);
+			userRefreshTimer.Start();
 		}
 
-		private int getSelectedServerId()
+		private int SelectedServerId
 		{
-			Server selectedServer = (Server)serverListView.SelectedItem;
-			if (selectedServer == null)
+			get
 			{
-                return -1;
-			}
-			else
-			{
-				return selectedServer.Id;
+				Server selectedServer = (Server)serverListView.SelectedItem;
+				if (selectedServer == null)
+				{
+					return -1;
+				}
+				else
+				{
+					return selectedServer.Id;
+				}
 			}
 		}
 
-		private async Task refreshJoinedServers()
+		private async Task RefreshJoinedServers()
 		{
             // Get all joined servers
             await chatInterface.GetServers();
@@ -57,11 +75,27 @@ namespace ChitChatClient.Windows
             }
         }
 
+		private async Task RefreshServerUsers()
+		{
+			await chatInterface.GetServerUsers(SelectedServerId);
+			userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
+
+			Server selectedServer = (Server)serverListView.SelectedItem;
+			int onlineUserCount = await serverService.GetOnlineUserCount(selectedServer.Id);
+
+			onlineUserCountTB.Text = "Online " + onlineUserCount.ToString();
+			totalUserCountTB.Text = "Total " + selectedServer.UserCount.ToString();
+		}
+
+		private async void RefreshServerUsersTimed(object? sender, EventArgs e)
+		{
+			await RefreshServerUsers();
+		}
+
         private async void Window_Loaded(object sender, RoutedEventArgs e)
         {
-            await refreshJoinedServers();
-            await chatInterface.GetServerUsers(getSelectedServerId());
-            userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
+            await RefreshJoinedServers();
+            await RefreshServerUsers();
         }
 
         private async void createServerButton_Click(object sender, RoutedEventArgs e)
@@ -69,7 +103,7 @@ namespace ChitChatClient.Windows
 			// Create a server
 			CreateServerWindow createServerWindow = new();
 			createServerWindow.ShowDialog();
-			await refreshJoinedServers();
+			await RefreshJoinedServers();
         }
 
 		private async void removeServerButton_Click(object sender, RoutedEventArgs e)
@@ -77,19 +111,19 @@ namespace ChitChatClient.Windows
 			// Remove a server
 			if (serverListView.SelectedIndex != -1)
 			{
-				bool result = await chatInterface.LeaveServer(getSelectedServerId());
+				bool result = await chatInterface.LeaveServer(SelectedServerId);
 
 				if (result)
 				{
 					MessageBox.Show("Server left successfully", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
-                    await refreshJoinedServers();
+                    await RefreshJoinedServers();
 				}
 				else
 				{
 					MessageBox.Show("Failed to leave server", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
 				}
 
-				await refreshJoinedServers();
+				await RefreshJoinedServers();
                 serverListView.SelectedIndex = 0;
 			}
         }
@@ -99,7 +133,7 @@ namespace ChitChatClient.Windows
 		{
 			ServerListWindow serverListWindow = new();
 			serverListWindow.ShowDialog();
-			await refreshJoinedServers();
+			await RefreshJoinedServers();
         }
 
 		private void serverInfoButton_Click(object sender, RoutedEventArgs e)
@@ -130,8 +164,13 @@ namespace ChitChatClient.Windows
 
 		private async void serverListView_SelectionChanged(object sender, SelectionChangedEventArgs e)
 		{
-			await chatInterface.GetServerUsers(getSelectedServerId());
+			await chatInterface.GetServerUsers(SelectedServerId);
             userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
         }
-    }
+
+		private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
+		{
+			userRefreshTimer?.Stop();
+		}
+	}
 }