+++ release_notes.md
@@ -1,9 +1,11 @@
 ## New features
 
 - Check for new version when starting the client and notify the user if a new version is available
+- Added on status to user list which refreshes every 5 seconds
 
 ## Changes
 
+- Increased chat window size to 1000x540
 
 ## Fixes